Adds TracingCachingHttpClientBuilder for httpclient-cache
This refactors types so that we can introduce identical functionality
from TracingHttpClientBuilder for TracingCachingHttpClientBuilder.

Fixes #431
